+++ gridsome/__tests__/__fixtures__/project-path-prefix/src/pages/About.vue
@@ -1,0 +1,5 @@
+<template>
+  <div>
+    about
+  </div>
+</template>