+++ gridsome/lib/server/middlewares/assets.js
@@ -46,6 +46,7 @@ module.exports = ({ context, config, queue }) => {
 
         if (!fs.existsSync(destPath)) {
           await worker.processImage({
+            backgroundColor: config.images.backgroundColor,
             minWidth: config.minProcessImageWidth,
             size: asset.size,
             filePath,